+++ lib/ReactViews/ExplorerWindow/Tabs/DataCatalogTab.jsx
@@ -49,12 +49,8 @@ class DataCatalogTab extends React.Component {
     return (
       <div className={Styles.root}>
         <Box fullHeight column>
-          <Box
-            css={`
-              height: ${showBreadcrumbs ? `calc(100% - 32px)` : `100%`};
-            `}
-          >
-            <Box className={Styles.dataExplorer} flex="1 1 40rem">
+          <Box fullHeight overflow="hidden">
+            <Box className={Styles.dataExplorer} styledWidth="40%">
               {/* ~TODO: Put this back once we add a MobX DataCatalogSearch Provider~ */}
               {/* TODO2: Implement a more generic MobX DataCatalogSearch */}
               {searchState.catalogSearchProvider && (
@@ -62,7 +58,6 @@ class DataCatalogTab extends React.Component {
                   searchText={searchState.catalogSearchText}
                   onSearchTextChanged={val => this.changeSearchText(val)}
                   onDoSearch={() => this.search()}
-                  onClear={() => this.props.viewState.showBreadcrumbs(false)}
                   placeholder={this.searchPlaceholder}
                   debounceDuration={
                     terria.catalogReferencesLoaded &&
@@ -81,7 +76,7 @@ class DataCatalogTab extends React.Component {
                 items={this.props.items}
               />
             </Box>
-            <Box flex="1 1 60rem">
+            <Box styledWidth="60%" wordBreak="break-all">
               <DataPreview
                 terria={terria}
                 viewState={this.props.viewState}